update command serial

+++ include/cavan/text.h
@@ -130,7 +130,7 @@ double text2size(const char *text, const char **last);
 double text2size_mb(const char *text);
 char *size2text_base(u64 size, char *buff, size_t buff_len);
 char *size2text(u64 size);
-u64 text2clock(const char *text);
+double text2clock(const char *text);
 double text2time_single(const char *text, const char **last);
 double text2time(const char *text, const char **last);
 int text2date(const char *text, struct tm *date, ...);
